Refactor: Flatten modules and remove dead code (Razor)
- Deleted `src/cli.rs` (dead code, duplicated in `src/tools/runner.rs`).
- Flattened `src/ast/` into `src/ast.rs`.
- Flattened `src/errors/` into `src/errors.rs`.
- Updated imports in `src/semantic/assembler.rs`.

This simplification reduces file count and directory depth, adhering to KISS and YAGNI principles. All tests pass.
